DR-172: made the sorting work with the new SubComponent
